moved main loop/reporting to separate module
